cube-dev-demo/sample_data/requirements.txt
Faker==19.11.0
pandas==2.0.3
snowflake-sqlalchemy==1.5.0
python-dotenv==1.0.0
jupyter==1.0.0
